dynamic pricing on initial page load , dynamic pricing on child infant adult change , added circular progress on loading , on startup selected date is made to date time now +cutoff hours , addon pricing added to final price

+++ travelerdubai/lib/tourdetails/touroption_data_layer/model/response/tour_option_dynamic_response.dart
@@ -1,3 +1,5 @@
+
+
 import 'package:json_annotation/json_annotation.dart';
 
 part 'tour_option_dynamic_response.g.dart';
@@ -18,9 +20,9 @@ class TourOptionDynamicDataResponse {
 
 @JsonSerializable()
 class ExtractedData {
-  int? addPriceAdult;
-  int? addPriceChildren;
-  int? additionalPriceInfant;
+  double? addPriceAdult;
+  double? addPriceChildren;
+  double? additionalPriceInfant;
 
   ExtractedData({this.addPriceAdult, this.addPriceChildren, this.additionalPriceInfant});
